[fix] コンポーネントの分離
