+++ src/components/modal/index.js
@@ -1,0 +1,3 @@
+import SearchFormModal from "./SearchFormModal";
+
+export { SearchFormModal };
